requirements.txt
mysql-connector-python==8.0.5
python-dateutil~=2.8.2
numpy~=1.23.5